Update Import.php
This was a fix i see J. did that fixed a problem for a client. I just had to do some digging as they were again having an issue, but this time was their data. so having found this. and while being a bit clueless on what it fixed, i thought I should put it here for others to review.