Update importer for 5.51
I've re-written this to work with 5.51 and
stripped out a lot of unnecessary code.

Note that while interacting with core import classes is not supported
this extension has done that over a number of years & has been
fairly heavily used by the community (although not by me) as the rapidity
with which issues get logged when it breaks shows. This update
still leaves it interacting with core in
a way that is unsupported by core but it gets it to be non-broken

Note the main features of this extension are
- Queue importing - this is now in core & the extension uses
  the core implementation. I have removed the ability to
  specify the batch size as that should be added as
  a feature to core rather than this extension

- Ability to import to any api

- The ability to toggle case-insensitive importing - this is
  likely not working at the moment and it will do as per core
  and always permit case-insensitive matching. It should be
  addressed in core if it is broken I think

- A bit of matching fields magic - I haven't tested this so I don't know
  the state of it as the moment or what the future should be for this logic

Going forwards I have yet to decide the future of this extension  - my current goal is
that it is not broken right now :-) I also think the ability to import to any entity
is valuable -but I don't know if this extension is the best way to deliver that